Added hide_context option to collection and tag admin
