+++ Admin/ContextAwareAdmin.php
@@ -1,0 +1,106 @@
+<?php
+
+/*
+ * This file is part of the Sonata package.
+ *
+ * (c) Thomas Rabaix <thomas.rabaix@sonata-project.org>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sonata\ClassificationBundle\Admin;
+
+use Sonata\AdminBundle\Admin\Admin;
+use Sonata\AdminBundle\Datagrid\DatagridMapper;
+use Sonata\ClassificationBundle\Entity\ContextManager;
+use Sonata\ClassificationBundle\Model\ContextInterface;
+use Sonata\ClassificationBundle\Model\ContextManagerInterface;
+
+abstract class ContextAwareAdmin extends Admin
+{
+    /**
+     * @var ContextManagerInterface
+     */
+    protected $contextManager;
+
+    /**
+     * @param string                  $code
+     * @param string                  $class
+     * @param string                  $baseControllerName
+     * @param ContextManagerInterface $contextManager
+     */
+    public function __construct($code, $class, $baseControllerName, ContextManagerInterface $contextManager)
+    {
+        parent::__construct($code, $class, $baseControllerName);
+
+        $this->contextManager = $contextManager;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getNewInstance()
+    {
+        $instance = parent::getNewInstance();
+
+        if ($contextId = $this->getPersistentParameter('context')) {
+            $context = $this->contextManager->find($contextId);
+
+            if (!$context) {
+                /** @var ContextInterface $context */
+                $context = $this->contextManager->create();
+                $context->setEnabled(true);
+                $context->setId($contextId);
+                $context->setName($contextId);
+
+                $this->contextManager->save($context);
+            }
+
+            $instance->setContext($context);
+        }
+
+        return $instance;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function configureDatagridFilters(DatagridMapper $datagridMapper)
+    {
+        $options = array();
+
+        if (1 === $this->getPersistentParameter('hide_context')) {
+            $options['disabled'] = true;
+        }
+
+        $datagridMapper
+            ->add('context', null, array(), null, $options)
+        ;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getPersistentParameters()
+    {
+        $parameters = array(
+            'context'      => '',
+            'hide_context' => $this->hasRequest() ? (int) $this->getRequest()->get('hide_context', 0) : 0,
+        );
+
+        if ($this->getSubject()) {
+            $parameters['context'] = $this->getSubject()->getContext() ? $this->getSubject()->getContext()->getId() : '';
+
+            return $parameters;
+        }
+
+        if ($this->hasRequest()) {
+            $parameters['context'] = $this->getRequest()->get('context');
+
+            return $parameters;
+        }
+
+        return $parameters;
+    }
+}
